middleware/requirements.txt
click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
itsdangerous==1.1.0
MarkupSafe==1.1.0
psycopg2-binary==2.7.4
setuptools==40.8.0
wheel==0.33.1
Werkzeug==0.14.1
requests==2.20.1
boto3==1.9.105
Flask-GraphQL==2.0.0
graphene==2.1.3

